settings page menu added and sidebar components' modified
