+++ src/app/(afterLogin)/settings/page.tsx
@@ -1,175 +1,70 @@
 "use client";
 
-import { Suspense, useState } from "react";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
-import { Avatar, AvatarFallback, AvatarImage } from "@/components/ui/avatar";
-import Navbar from "@/components/Navbar";
+import React from "react";
+import Link from "next/link";
 import { SidebarTrigger } from "@/components/ui/sidebar";
-import { authClient } from "@/lib/auth-client";
-import { addPasskey } from "@/lib/auth-actions";
+import {
+  Shield,
+  User,
+  Bell,
+  Lock,
+  HelpCircle,
+  Globe,
+  DollarSign,
+  Star,
+  Users,
+} from "lucide-react";
 
-export default function SettingsPage() {
-  const { data: session } = authClient.useSession();
+// Sidebar menu items
+const settingsMenu = [
+  { title: "Your account", url: "#", icon: User },
+  { title: "Monetization", url: "#", icon: DollarSign },
+  { title: "Premium", url: "#", icon: Star },
+  { title: "Creator Subscriptions", url: "#", icon: Users },
+  { title: "Security and account access", url: "#", icon: Shield },
+  { title: "Privacy and safety", url: "#", icon: Lock },
+  { title: "Notifications", url: "#", icon: Bell },
+  {
+    title: "Accessibility, display, and languages",
+    url: "#",
+    icon: Globe,
+  },
+  { title: "Additional resources", url: "#", icon: HelpCircle },
+  { title: "Help Center", url: "#", icon: HelpCircle },
+];
 
-  const [isEditing, setIsEditing] = useState(false);
+export default function Settings() {
+
+
 
   return (
-    <div className="w-full bg-gradient-to-r from-blue-500 to-purple-500">
-      {/* <Navbar /> */}
-      <SidebarTrigger />
-      <div className="px-6 py-16 max-w-4xl mx-auto">
-        {/* welcome - Header Section */}
-        <div className="flex justify-between items-center">
-          <h2 className="text-2xl font-bold text-blue-900">
-            Welcome, {session?.user?.username ?? "User!"}
-          </h2>
-        </div>
+    <div className="w-full min-h-screen flex flex-col sm:flex-row">
+      {/* Mobile Sidebar Toggle */}
+      <div className="p-4 sm:hidden">
+        <SidebarTrigger className="w-10 h-10 " />
+      </div>
 
-        <p className="text-blue-900 text-sm"></p>
+      {/* Sidebar */}
+      <div className="w-full border-r border-gray-800 p-8 flex flex-col justify-center sm:justify-start">
 
-        <div className="bg-blue-200 rounded-lg shadow-lg p-6 mt-4">
-          {/* Profile Card */}
-          <div className="flex items-center justify-between">
-            {/* Left Section: Avatar & Name */}
-            <div className="flex gap-6 items-center">
-              <Suspense fallback = {<AvatarFallback>{session?.user.name.charAt(0)}</AvatarFallback>} >
-              <Avatar className="w-16 h-16">
-                <AvatarImage src={ session?.user?.image ?? undefined} alt={session?.user?.username ?? ""} />
-  
-              </Avatar>
-              </Suspense>
-              <div>
-                <h3 className="text-lg font-semibold text-gray-900">
-                  {session?.user?.name}
-                </h3>
-                <p className="text-gray-700 text-sm">{session?.user?.email}</p>
-              </div>
-            </div>
+        <h2 className="text-xl font-semibold mb-6">Settings</h2>
+        <ul className="space-y-1">
+          {settingsMenu.map((item, index) => {
+            const Icon = item.icon;
 
-            {/* Right Section: Button */}
-            <Button variant="secondary"
-              className="bg-blue-500 text-white"
-              onClick={() => setIsEditing(!isEditing)}
-            >
-              {isEditing ? "Save" : "Edit"}
-            </Button>
-          </div>
-
-          {/* Profile Form */}
-          <div className="grid grid-cols-2 gap-4 mt-6">
-            <div>
-              <label className="text-sm text-gray-600">Full Name</label>
-              <Input className="text-black"
-                disabled={!isEditing}
-                placeholder="Your First Name"
-                defaultValue={session?.user?.name}
-              />
-            </div>
-            {/* <div className="flex ">
-
-              <Button
-                className="bg-blue-500 text-white"
-                onClick={() => setIsEditing(!isEditing)}
-              >
-                {isEditing ? "Save" : "Edit"}
-              </Button>
-            </div> */}
-
-            <div>
-              <label className="text-sm text-gray-600">Nick Name</label>
-              <Input className="text-black"
-                disabled={!isEditing}
-                placeholder="Your Nickname"
-                defaultValue={session?.user?.username ?? ""}
-              />
-            </div>
-
-            <div>
-              <label className="text-sm text-gray-600">Gender</label>
-              <select
-                className="border rounded-md p-2 w-full text-black"
-                disabled={!isEditing}
-                defaultValue={"Male"}
-              >
-                <option>Your Gender</option>
-                <option>Male</option>
-                <option>Female</option>
-                <option>Other</option>
-                <option>Prefer not to say</option>
-              </select>
-            </div>
-
-            <div>
-              <label className="text-sm text-gray-600">Country</label>
-              <select
-                className="border rounded-md p-2 w-full text-black"
-                disabled={!isEditing}
-                defaultValue={"India"}
-              >
-                <option>Your Country</option>
-                <option>India</option>
-                <option>USA</option>
-                <option>UK</option>
-              </select>
-            </div>
-
-            <div>
-              <label className="text-sm text-gray-600">Language</label>
-              <select
-                className="border rounded-md p-2 w-full text-black"
-                disabled={!isEditing}
-                defaultValue={"English"}
-              >
-                <option>Your Language</option>
-                <option>English</option>
-                <option>Bangla</option>
-              </select>
-            </div>
-          </div>
-
-          {/* Email Address Section */}
-          <div className="mt-6">
-            <h3 className="text-lg font-semibold">My Email Address</h3>
-            <div className="flex items-center gap-4 bg-gray-100 p-3 rounded-md">
-              <span className="bg-blue-500 text-white rounded-full p-2">
-                📧
-              </span>
-              <div>
-                <p className="text-sm text-black font-medium">{session?.user?.email}</p>
-                <p className="text-xs text-gray-500">1 month ago</p>
-              </div>
-            </div>
-
-            <Button variant="secondary" className="bg-blue-500 text-white mt-4">
-              + Add Email Address
-            </Button>
-
-            <Button onClick={addPasskey} variant="secondary" className="m-2 mt-4">
-              + Add PassKey
-            </Button>
-          </div>
-        </div>
-
-        {/* Other Details*/}
-        <div className="bg-blue-200 rounded-lg shadow-lg p-6 mt-4">
-          <div className="grid grid-cols-2 gap-4 mt-6">
-            <div>
-              <label className="text-sm text-gray-600">Total Friends</label>
-              <Input className="text-black" disabled={true} placeholder="Nil" defaultValue={"120"} />
-            </div>
-
-            <div>
-              <label className="text-sm text-gray-600">Total Posts</label>
-              <Input className="text-black" disabled={true} placeholder="Nil" defaultValue={"56"} />
-            </div>
-
-            <div>
-              <label className="text-sm text-gray-600">Profile Reach</label>
-              <Input className="text-black" disabled={true} placeholder="Nil" defaultValue={"1.2k"} />
-            </div>
-          </div>
-        </div>
+            return (
+              <li key={index}>
+                <Link
+                  href={item.url}
+                  className={`flex items-center gap-3 py-2 px-3 rounded-md text-sm font-medium transition-colors hover:bg-gray-500`}
+                >
+                  <Icon className="w-5 h-5" />
+                  <span>{item.title}</span>
+                </Link>
+              </li>
+            );
+          })}
+        </ul>
       </div>
     </div>
   );
